refactor: response.tsのインポートパスを修正
