+++ src/application/api/article/getArticles.test.ts
@@ -1,7 +1,7 @@
 import getRdbClient, { RdbClient } from '@/infrastructure/rdb';
 import app from '../../server';
 import TEST_ENV from '@/test/env';
-import { ArticleListResponse } from './types/response';
+import { ArticleListResponse } from './response';
 
 type GetArticlesTestCase = {
   name: string;
